# Soundex: Performance Optimizations