--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/xrash/smetrics
+
+go 1.21.1