--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,0 @@
-module github.com/xrash/smetrics
-
-go 1.21.1