--- a/tests/soundex_test.go
+++ b/tests/soundex_test.go
@@ -6,24 +6,6 @@ import (
 	"testing"
 )
 
-//func BenchmarkSoundex(b *testing.B) {
-//	fd, err := os.Open("testdata/english_names.txt")
-//	if err != nil {
-//		b.Fatal(err)
-//	}
-//	defer fd.Close()
-//
-//	scanner := bufio.NewScanner(fd)
-//	for scanner.Scan() {
-//		s := smetrics.Soundex(scanner.Text())
-//
-//		// Trivially use the result to avoid compiler optimizations.
-//		if len(s) != 4 {
-//			b.Fail()
-//		}
-//	}
-//}
-
 func TestSoundex(t *testing.T) {
 	cases := []soundexcase{
 		{"Euler", "E460"},