Hello!

I have optimized the soundex function's cpu performance by a little over 10x.

| Current | This PR |
| --- | --- |
| `0.3875 ns/op` | `0.03767 ns/op` |

At your convenience and discretion, please pull the following changes:

Project:
1.  Created a go.mod file to allow use of `go test` at the project level

Soundex:
1. Removed map and replaced with a `switch case` statement for performance, which avoids allocations and map access latency.
2. Switched all letter variables to bytes, since this function only supports ASCII anyway.
3. Removed all function calls in the inner loop. (`toUpper` was especially slow)
4. Used a stringbuilder so that only a single allocation is made per iteration.
5. I have included the bechmark code I used. This was tested against 500k randomly generated english names. I am not including the test set, so I have commented out the benchmark.