@xrash 2023-12-13
Hi Michael, thanks for your time and effort contributing to smetrics!

I have a few points to make:

1) Can we let go.mod to be introduced in another PR? That's something I've been postponing for ever at this point and shall take a look at it. Also I'm sure others are using outdated versions of Go and setting the minimum version of the package to 1.21.1 wouldn't be ideal.

2) The current code doesn't check for the validity of the first letter; yours, on the other hand, does, and returns `0000` if it's not valid. I appreciate the attention to this detail, but, in order to avoid changing behavior of code already in use, and because `0000` doesn't look like a valid soundex code after all, I'd rather keep the current behavior. In a future version, if we want to handle this situation differently, an interesting approach could be to simply skip anything outside a-z and consider the first valid letter it finds. For reference, this is what the soundex implementation in postgres' fuzzystrmatch extension does. 

3) Regarding the commented code, I'd rather not include it, even more not so in the test file. I can imagine a "benchmark" directory at the root of the project with stuff like this and the data used in benchmarking could be hosted somewhere else. But I'm not even convinced of the value of having benchmark code in the repo as of now. What do you think?

Thanks again!