@michaelclarkcuadrado 2023-12-13
I believe I've fixed up those three points:

1. Removed go.mod
2. Removed the first letter check. Just like the existing code, I capitalize it if it is lowercase ascii, but otherwise do no validation.
3. I've removed the commented benchmark.